Refactored code, added extIPA support, fixed various bugs
