+++ IPA Keyboard Extension/Base.lproj/Localizable.strings
@@ -15,6 +15,7 @@
 "Suprasegmentals" = "Suprasegmentals";
 "TonesAndWordAccents" = "Tones and Word Accents";
 "Diacritics" = "Diacritics";
+"ExtIPA" = "extIPA";
 
 "SpaceBarText" = "space";
