+++ IPA Keyboard Extension/eo.lproj/Localizable.strings
@@ -15,6 +15,7 @@
 "Suprasegmentals" = "Supersegmentaĵoj";
 "TonesAndWordAccents" = "Tonoj kaj akcentoj";
 "Diacritics" = "Diakritiloj";
+"ExtIPA" = "Aldonoj al la IFA";
 
 "SpaceBarText" = "spaceto";
