+++ IPA Keyboard Extension/ja.lproj/Localizable.strings
@@ -15,7 +15,8 @@
 "Suprasegmentals" = "超分節要素";
 "TonesAndWordAccents" = "声調記号とアクセント";
 "Diacritics" = "補助記号";
+"ExtIPA" = "拡張IPA";
 
 "SpaceBarText" = "空白";
 
-"IPAWikipediaURL" = "https://ja.wikipedia.org/wiki/国際音声記号";
+"IPAWikipediaURL" = "https://ja.wikipedia.org/wiki/%E5%9B%BD%E9%9A%9B%E9%9F%B3%E5%A3%B0%E8%A8%98%E5%8F%B7"; // 国際音声記号
