+++ IPA Keyboard Extension/zh-Hans.lproj/Localizable.strings
@@ -15,7 +15,8 @@
 "Suprasegmentals" = "超声段成分";
 "TonesAndWordAccents" = "声调";
 "Diacritics" = "变音符号";
+"ExtIPA" = "国际音标扩展";
 
 "SpaceBarText" = "空格";
 
-"IPAWikipediaURL" = "https://baike.baidu.com/item/国际音标";
+"IPAWikipediaURL" = "https://baike.baidu.com/item/%E5%9B%BD%E9%99%85%E9%9F%B3%E6%A0%87";　// 国际音标
