+++ IPA Keyboard Extension/zh-Hant.lproj/Localizable.strings
@@ -15,7 +15,8 @@
 "Suprasegmentals" = "超音段成分";
 "TonesAndWordAccents" = "聲調";
 "Diacritics" = "變音符號";
+"ExtIPA" = "國際音標擴展";
 
 "SpaceBarText" = "空格";
 
-"IPAWikipediaURL" = "https://zh.wikipedia.org/wiki/國際音標";
+"IPAWikipediaURL" = "https://zh.wikipedia.org/wiki/%E5%9C%8B%E9%9A%9B%E9%9F%B3%E6%A8%99"; // 國際音標
