+++ IPA Keyboard Shared/GlobalSymbols.swift
@@ -1,0 +1,18 @@
+//
+//  GlobalSymbols.swift
+//  IPA Keyboard
+//
+//  Created by Vincent C. on 2019/7/26.
+//  Copyright © 2019 Vince14Genius. All rights reserved.
+//
+
+import Foundation
+
+enum GlobalSymbols {
+    
+    static let squareBrackets = "[]"
+    static let forwardSlashes = "//"
+    static let tilde = "~"
+    static let dottedCircle: Character = "◌"
+    
+}
