+++ IPA Keyboard.xcodeproj/xcuserdata/vince14genius.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -14,6 +14,11 @@
 			<key>orderHint</key>
 			<integer>0</integer>
 		</dict>
+		<key>extIPA Keyboard Extension.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>2</integer>
+		</dict>
 	</dict>
 	<key>SuppressBuildableAutocreation</key>
 	<dict>
