+++ extIPA Keyboard Extension/Localization/ja.lproj/Localizable.strings
@@ -1,0 +1,9 @@
+/* 
+  Localizable.strings
+  IPA Keyboard
+
+  Created by Vincent C. on 4/28/20.
+  Copyright © 2020 Vince14Genius. All rights reserved.
+*/
+
+"CFBundleDisplayName" = "拡張IPA";
